requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
cupy-cuda11x>=11.0.0; sys_platform != "win32"  # GPU acceleration (optional, Linux/Mac)

